+++ embark/workers/orchestrator.py
@@ -1,3 +1,4 @@
+import logging
 from uuid import UUID
 from typing import Dict, List
 from collections import deque
@@ -7,6 +8,8 @@ from redis import Redis
 from django.conf import settings
 
 from workers.models import Worker, OrchestratorState
+
+logger = logging.getLogger(__name__)
 
 
 REDIS_CLIENT = Redis(host=settings.REDIS_HOST, port=settings.REDIS_PORT)
@@ -49,6 +52,25 @@ class Orchestrator:
         with REDIS_CLIENT.lock(LOCK_KEY, LOCK_TIMEOUT):
             self._sync_orchestrator_state()
             return self.free_workers
+
+    def _assign_tasks(self):
+        """
+        Assign tasks to free workers as long as there are both tasks and free workers available.
+        """
+        if self.tasks and self.free_workers:
+            next_task = self.tasks.popleft()
+            free_worker = next(iter(self.free_workers.values()))
+            self._assign_worker(free_worker, next_task)
+            self._assign_tasks()
+
+    def assign_tasks(self):
+        """
+        Trigger the orchestrator to assign tasks to free workers.
+        """
+        with REDIS_CLIENT.lock(LOCK_KEY, LOCK_TIMEOUT):
+            self._sync_orchestrator_state()
+            self._assign_tasks()
+            self._update_orchestrator_state()
 
     def get_busy_workers(self) -> Dict[str, Worker]:
         with REDIS_CLIENT.lock(LOCK_KEY, LOCK_TIMEOUT):
@@ -96,34 +118,24 @@ class Orchestrator:
             self._sync_orchestrator_state()
             return self._get_worker_info(worker_ips)
 
-    def _assign_task(self, task: OrchestratorTask):
+    def _queue_task(self, task: OrchestratorTask):
         """
-        Assign a task to a worker. If no workers are free, the task is queued.
-        If there are free workers and no tasks in the queue, the task is assigned immediately.
-        Otherwise, assigns queued tasks to free workers until no more free workers are available.
+        Add a task to the task queue and trigger task assignment.
 
         :param task: The task to be assigned
         """
-        if not self.free_workers:
-            self.tasks.append(task)
-        else:
-            free_worker = next(iter(self.free_workers.values()))
-            if not self.tasks:
-                self._assign_worker(free_worker, task)
-            else:
-                queued_task = self.tasks.popleft()
-                self._assign_worker(free_worker, queued_task)
-                self._assign_task(task)
+        self.tasks.append(task)
+        self._assign_tasks()
 
-    def assign_task(self, task: OrchestratorTask):
+    def queue_task(self, task: OrchestratorTask):
         """
-        Checks the lock and calls `_assign_task`
+        Checks the lock and calls `_queue_task`
 
         :param task: The task to be assigned
         """
         with REDIS_CLIENT.lock(LOCK_KEY, LOCK_TIMEOUT):
             self._sync_orchestrator_state()
-            self._assign_task(task)
+            self._queue_task(task)
             self._update_orchestrator_state()
 
     def _assign_worker(self, worker: Worker, task: OrchestratorTask):
@@ -199,6 +211,8 @@ class Orchestrator:
             raise ValueError(f"Worker with IP {worker.ip_address} already exists.")
         self.free_workers[worker.ip_address] = worker
 
+        logger.info("Worker: %s added to orchestrator", worker.name)
+
     def add_worker(self, worker: Worker):
         """
         Add a new worker to the orchestrator. The worker is added to the free workers list.
@@ -211,30 +225,37 @@ class Orchestrator:
             self._add_worker(worker)
             self._update_orchestrator_state()
 
-    def _remove_worker(self, worker: Worker):
+    def _remove_worker(self, worker: Worker, check: bool = True):
         """
         Remove a worker from the orchestrator. The worker is removed from either the free or busy workers list.
 
         :param worker: The worker to be removed
-        :raises ValueError: If the worker does not exist in the orchestrator
+        :param check: If True, throws an error
+        :raises ValueError: If the worker does not exist in the orchestrator and check = True
         """
         if worker.ip_address in self.free_workers:
             del self.free_workers[worker.ip_address]
         elif worker.ip_address in self.busy_workers:
             del self.busy_workers[worker.ip_address]
         else:
-            raise ValueError(f"Worker with IP {worker.ip_address} does not exist.")
+            if check:
+                raise ValueError(f"Worker with IP {worker.ip_address} does not exist.")
 
-    def remove_worker(self, worker: Worker):
+            return
+
+        logger.info("Worker: %s removed from orchestrator", worker.name)
+
+    def remove_worker(self, worker: Worker, check: bool = True):
         """
         Remove a worker from the orchestrator.
 
         :param worker: The worker to be removed
-        :raises ValueError: If the worker does not exist in the orchestrator
+        :param check: If True, throws an error
+        :raises ValueError: If the worker does not exist in the orchestrator and check = True
         """
         with REDIS_CLIENT.lock(LOCK_KEY, LOCK_TIMEOUT):
             self._sync_orchestrator_state()
-            self._remove_worker(worker)
+            self._remove_worker(worker, check)
             self._update_orchestrator_state()
 
     def _get_orchestrator_state(self):